feat(dm): Expose Cli as a library
This will make it easier to build porcelain programs that combine the functionality of existing programs without creating more code paths that have to be tested and maintained.

`docs/commit-messages.md`:
- Register `dm` as a scope because every crate should be on the list .
