+++ Cargo.toml
@@ -36,6 +36,7 @@ rs4a-bin-utils = {path = "crates/bin-utils" }
 rs4a-dut = { version = "0.1", path = "crates/dut" }
 rs4a-vapix = { path = "crates/vapix" }
 rs4a-cassette-testing = { path = "crates/cassette-testing" }
+rs4a-device-manager = { path = "crates/device-manager" }
 rs4a-vlt = { path = "crates/vlt" }
 
 [workspace.package]
