+++ crates/device-manager/Cargo.toml
@@ -1,8 +1,12 @@
 [package]
-name = "device-manager"
+name = "rs4a-device-manager"
 version = "0.1.0"
 edition.workspace = true
 license = "MIT"
+
+[[bin]]
+name = "device-manager"
+path = "src/main.rs"
 
 [dependencies]
 anyhow = { workspace = true }
