+++ crates/device-manager/src/commands/reinit.rs
@@ -3,14 +3,14 @@ use crate::Netloc;
 #[derive(Clone, Debug, clap::Args)]
 pub struct ReinitCommand {
     #[command(flatten)]
-    netloc: Netloc,
+    pub netloc: Netloc,
 }
 
 impl ReinitCommand {
-    pub async fn exec(self) -> anyhow::Result<()> {
+    pub async fn exec(self) -> anyhow::Result<String> {
         super::init::require_root_user(&self.netloc)?;
         super::restore::restore(&self.netloc).await?;
         super::init::initialize(&self.netloc).await?;
-        Ok(())
+        Ok(String::new())
     }
 }
